refactor(api): move weather controller and service to feature-based folder structure
- Renamed and relocated weather_controller.ts → controllers/
- Renamed and relocated weather_service.ts → services/
- Updated import paths accordingly
