+++ functions/src/services/weather_service.ts
